+++ src/complex-array-base.ts
@@ -1,5 +1,6 @@
-import { QueryList, SimpleChanges, SimpleChange } from '@angular/core';
+import { QueryList, SimpleChanges, SimpleChange, EmbeddedViewRef } from '@angular/core';
 import { getValue } from '@syncfusion/ej2-base/util';
+import { clearTemplate } from './util';
 
 /**
  * Complex Array Base module
@@ -14,6 +15,7 @@ interface Tag {
     hasChanges: boolean;
     getProperties: Function;
     isInitChanges: boolean;
+    clearTemplate?: (args: string[]) => void;
 }
 
 export class ComplexBase<T> {
@@ -23,13 +25,22 @@ export class ComplexBase<T> {
     public property?: string;
     public tags?: string[] = [];
     private tagObjects?: { name: string, instance: Tag }[] = [];
-
+    private registeredTemplate: { [key: string]: EmbeddedViewRef<Object>[] };
     public ngOnInit(): void {
+        this.registeredTemplate = {};
         for (let tag of this.tags) {
             let objInstance: Tag = getValue('child' + tag.substring(0, 1).toUpperCase() + tag.substring(1), this);
             if (objInstance) {
                 this.tagObjects.push({ instance: objInstance, name: tag });
             }
+        }
+        let templateProperties: string[] = Object.keys(this);
+        templateProperties = templateProperties.filter((val: string): boolean => {
+            return /Ref$/i.test(val);
+        });
+        for (let tempName of templateProperties) {
+            let propName: string = tempName.replace('Ref', '');
+            this.propCollection[propName] = getValue(propName, this);
         }
     }
 
@@ -40,12 +51,30 @@ export class ComplexBase<T> {
         }
         this.hasChanges = true;
     }
-
+    public clearTemplate(templateNames: string[]): void {
+        clearTemplate(this, templateNames);
+    };
     public getProperties(): { [key: string]: Object } {
         for (let tagObject of this.tagObjects) {
             this.propCollection[tagObject.name] = tagObject.instance.getProperties();
         }
         return this.propCollection;
+    }
+
+    public isChanged(): boolean {
+        let result: boolean = this.hasChanges;
+        for (let item of this.tagObjects) {
+            result = result || item.instance.hasChanges;
+        }
+        return result;
+    }
+
+    public ngAfterContentChecked(): void {
+        this.hasChanges = this.isChanged();
+        let templateProperties: string[] = Object.keys(this);
+        templateProperties = templateProperties.filter((val: string) => {
+            return /Ref$/i.test(val);
+        });
     }
 
     public ngAfterViewChecked(): void {
@@ -94,6 +123,14 @@ export class ArrayBase<T> {
             result = result || item.hasChanges;
         }
         return !!this.list.length && result;
+    }
+
+    public clearTemplate(templateNames: string[]): void {
+        for (let item of this.list) {
+            item.clearTemplate(templateNames && templateNames.map((val: string): string => {
+                return new RegExp(this.propertyName).test(val) ? val.replace(this.propertyName + '.', '') : val;
+            }));
+        }
     }
 
     public ngAfterContentChecked(): void {