+++ src/util.ts
@@ -10,19 +10,61 @@ export function applyMixins(derivedClass: any, baseClass: any[]): void {
         });
     });
 }
+
+/**
+ * @private 
+ */
+export function clearTemplate(_this: any, templateNames?: string[]): void {
+    let regTemplates: string[] = Object.keys(_this.registeredTemplate);
+    if (regTemplates.length) {
+        let regProperties: string[] = templateNames && templateNames.filter(
+            (val: string) => {
+                return (/\./g.test(val) ? false : true);
+            });
+        for (let registeredTemplate of (regProperties && regProperties || regTemplates)) {
+            for (let rt of _this.registeredTemplate[registeredTemplate]) {
+                rt.destroy();
+            }
+            delete _this.registeredTemplate[registeredTemplate];
+        }
+    }
+    for (let tagObject of _this.tagObjects) {
+        if (tagObject.instance) {
+            tagObject.instance.clearTemplate((templateNames && templateNames.filter(
+                (val: string) => {
+                    return (new RegExp(tagObject.name).test(val) ? true : false);
+                })));
+        }
+    }
+}
+
+/**
+ * To set value for the nameSpace in desired object.
+ * @param {string} nameSpace - String value to the get the inner object
+ * @param {any} value - Value that you need to set.
+ * @param {any} obj - Object to get the inner object value.
+ * @return {void}
+ * @private
+ */
+export function setValue(nameSpace: string, value: any, object: any): any {
+    let keys: string[] = nameSpace.split('.');
+    let fromObj: any = object || {};
+    for (let i:number = 0; i < keys.length; i++) {
+        let key: string = keys[i];
+        if (i + 1 === keys.length) {
+            fromObj[key] = value === undefined ? {} : value;
+        } else if (fromObj[key] === undefined) {
+            fromObj[key] = {};
+        }
+        fromObj = fromObj[key];
+    }
+    return fromObj;
+}
+
 /* tslint:enable */
 
 
-export function getPropArray(obj: PropertyDetails[], prefixKey: string = ''): string[] {
-    let properties: string[] = [];
-    for (let prop of obj) {
-        let typeObj: PropertyCollectionInfo = (prop.type as
-            { prototype: { propList: PropertyCollectionInfo } }).prototype.propList;
-        let resObj: string[] = getPropArray(typeObj.complexProps, '');
-        properties.push(prop.propertyName);
-    }
-    return properties;
-}
+
 
 export interface PropertyCollectionInfo {
     props: PropertyDetails[];